refactor(tests): reorganize into semantically meaningful test files
Replace test_main.py, test_v3.py, and test_coverage.py with:
- test_cfg.py       — Config class tests
- test_get_emulator.py — pre-existing failing download tests (excluded by -k filter)
- test_inputs.py    — input normalization and formatting tests
- test_properties.py — emulator properties and accessor tests
- test_outputs.py   — output dataclass and error statistics tests
- test_emulator.py  — emulator construction and prediction tests
- test_2d_ps.py     — 2D PS / score model / multi-GPU tests
- test_accuracy.py  — database accuracy comparison tests
