+++ pyproject.toml
@@ -58,6 +58,7 @@ dev = [
     "mypy>=0.930",
     "pre-commit>=2.16.0",
     "pre-commit-hooks>=4.1.0",
+    "pytest-cov>=7.1.0",
     "pyupgrade>=2.29.1",
     "ruff>=0.15.11",
     { include-group = "test" },
